move util/editor/index.js to components/editor/constants.js
