+++ apps/dashboard/src/components/editor/index.js
@@ -23,7 +23,7 @@ import Toolbar from './toolbar';
 import UnpublishedChangesNotice from './unpublished-changes-notice';
 import { useEditorContent } from './use-editor-content';
 import { useStylesheet } from '@crowdsignal/hooks';
-import { EDITOR_VIEW_AUTO, EDITOR_VIEW_DRAFT } from '../../util/editor';
+import { EDITOR_VIEW_AUTO, EDITOR_VIEW_DRAFT } from './constants';
 
 const Editor = ( { projectId, theme = 'leven' } ) => {
 	const [ forceDraft, setForceDraft ] = useState( false );
